Start convert all class variables to pointers.
inconsistent state- only Motor is ready

+++ ConsoleApplication1/ArduinoStub.cpp
@@ -41,12 +41,12 @@ void pinMode(int, int)
 
 void digitalWrite(int pin, int value)
 {
-	ArduinoPinsRegistry::arduinoPinsRegistry.setPinValue(pin, (float)(value));
+	ArduinoPinsRegistry::arduinoPinsRegistry.setPinValue(pin, value);
 }
 
 void analogWrite(int pin, int value)
 {
-	ArduinoPinsRegistry::arduinoPinsRegistry.setPinValue(pin, (float)(value));
+	ArduinoPinsRegistry::arduinoPinsRegistry.setPinValue(pin, value);
 }
 
 bool digitalRead(int)
